CORE-5120: Move and aggregate `FakeSecurityManager`
