+++ libs/http-rpc/http-rpc-client/build.gradle
@@ -39,7 +39,6 @@ dependencies {
 
     integrationTestImplementation project(":libs:http-rpc:http-rpc-server")
     integrationTestImplementation project(":libs:http-rpc:http-rpc-server-impl")
-    integrationTestImplementation project(":libs:http-rpc:http-rpc-security-read")
     integrationTestImplementation project(':libs:http-rpc:http-rpc-stubs-impl')
     integrationTestImplementation project(':libs:utilities')
     integrationTestImplementation project(':libs:http-rpc:ssl-cert-read')
