+++ libs/http-rpc/http-rpc-server-impl/src/integrationTest/kotlin/net/corda/httprpc/server/impl/HttpRpcServerTestBase.kt
@@ -2,6 +2,7 @@ package net.corda.httprpc.server.impl
 
 import net.corda.httprpc.server.HttpRpcServer
 import net.corda.httprpc.server.config.models.HttpRpcContext
+import net.corda.httprpc.test.utls.FakeSecurityManager
 import net.corda.httprpc.test.utls.TestHttpClient
 import java.net.ServerSocket
