+++ libs/http-rpc/http-rpc-test-common/build.gradle
@@ -21,6 +21,7 @@ dependencies {
     implementation project(":libs:http-rpc:http-rpc")
     implementation project(":libs:http-rpc:http-rpc-common")
     implementation project(":libs:http-rpc:http-rpc-tools")
+    implementation project(":libs:http-rpc:http-rpc-security-read")
     implementation project(":libs:serialization:serialization-amqp")
     implementation project(":libs:lifecycle:lifecycle")
 }
